🌹 Use usize for num_bits

+++ src/bit_buffer.rs
@@ -12,12 +12,12 @@ impl<'a> BitBuffer<'a> {
         self.buffer.len()
     }
 
-    pub fn write_lte25(&mut self, bit_values: u32, num_bits: u32) {
+    pub fn write_lte25(&mut self, bit_values: u32, num_bits: usize) {
         debug_assert!(num_bits > 0 && num_bits <= 25);
-        debug_assert!(self.position + (num_bits as usize) <= self.buffer.len() * 8);
+        debug_assert!(self.position + num_bits <= self.buffer.len() * 8);
 
         let target = unsafe { self.buffer.as_mut_ptr().add(self.position >> 3) as *mut u32 };
-        let shift = 32 - ((self.position as u32) & 7) - num_bits;
+        let shift = 32 - (self.position & 7) - num_bits;
 
         let curr = u32::from_be(unsafe { core::ptr::read_unaligned(target) });
 
@@ -26,12 +26,12 @@ impl<'a> BitBuffer<'a> {
 
         unsafe { core::ptr::write_unaligned(target, main.to_be()); }
 
-        self.position += num_bits as usize;
+        self.position += num_bits;
     }
 
-    pub fn write(&mut self, bit_values: u32, num_bits: u32) {
+    pub fn write(&mut self, bit_values: u32, num_bits: usize) {
         debug_assert!(num_bits > 0 && num_bits <= 32);
-        debug_assert!(self.position + (num_bits as usize) <= self.buffer.len() * 8);
+        debug_assert!(self.position + num_bits <= self.buffer.len() * 8);
 
         let target = unsafe { self.buffer.as_mut_ptr().add(self.position >> 3) as *mut u32 };
         let shift = (32 - ((self.position as i32) & 7) - (num_bits as i32)) as i32;
@@ -52,12 +52,12 @@ impl<'a> BitBuffer<'a> {
             unsafe { core::ptr::write_unaligned(target, main.to_be()); }
         }
 
-        self.position += num_bits as usize;
+        self.position += num_bits;
     }
 
     /// Align bit buffer to next byte boundary, writing zeros if requested
     pub fn byte_align(&mut self) {
-        let bit = (self.position & 7) as u32;
+        let bit = self.position & 7;
 
         if bit == 0 { return; }
